(TOOLCHAIN): move CargoProjectState class
